inhouse/requirements-train.txt
bdrk[model-monitoring]==0.6.0
torchvision==0.9.0
opencv-python==4.5.1.48
pandas==1.2.3
scikit-image==0.18.1
scikit-learn==0.24.1
gcsfs==0.7.2
google-cloud-storage==1.36.2
s3fs==0.5.2
fastparquet==0.5.0
pretrainedmodels==0.7.4
